Externalised, funcational (unit) connector script
